requirements.txt
fastapi
uvicorn[standard]
langgraph
langchain-core
python-dotenv
google-generativeai
yt_dlp
opencv-python
pillow
numpy
requests
